capybaras added
capybaras for first level positioned appropriately. the function in js allows them to show up one by one when you press the "Answer 2" button over and over again. again, you have to do this before pressing the other buttons as it could mess it up. pressing "answer 1" after this allows both the bushes and the capybaras to clear and gives way for level 2.